# [SYSTEMDS-3484] Frame Append Optimizations