--- a/src/main/java/org/apache/sysds/runtime/frame/data/columns/Array.java
+++ b/src/main/java/org/apache/sysds/runtime/frame/data/columns/Array.java
@@ -202,6 +202,17 @@ public abstract class Array<T> implements Writable {
 	public abstract void append(T value);
 
 	/**
+	 * append other array, if the other array is fitting in current allocated size use that allocated size, otherwise
+	 * allocate new array to combine the other with this.
+	 * 
+	 * This method should use the set range function, and should be preferred over the append single values.
+	 * 
+	 * @param other The other array of same type to append to this.
+	 * @return The combined arrays.
+	 */
+	public abstract Array<T> append(Array<T> other);
+
+	/**
 	 * Slice out the sub range and return new array with the specified type.
 	 * 
 	 * If the conversion fails fallback to normal slice