--- a/src/main/java/org/apache/sysds/runtime/frame/data/columns/DoubleArray.java
+++ b/src/main/java/org/apache/sysds/runtime/frame/data/columns/DoubleArray.java
@@ -143,8 +143,11 @@ public class DoubleArray extends Array<Double> {
 
 	@Override
 	public void reset(int size) {
-		if(_data.length < size)
+		if(_data.length < size || _data.length > 2 * size)
 			_data = new double[size];
+		else
+			for(int i = 0; i < size; i++)
+				_data[i] = 0;
 		_size = size;
 	}