--- a/src/main/java/org/apache/sysds/runtime/frame/data/columns/StringArray.java
+++ b/src/main/java/org/apache/sysds/runtime/frame/data/columns/StringArray.java
@@ -107,6 +107,15 @@ public class StringArray extends Array<String> {
 	}
 
 	@Override
+	public StringArray append(Array<String> other) {
+		final int endSize = this._size + other.size();
+		final String[] ret = new String[endSize];
+		System.arraycopy(_data, 0, ret, 0, this._size);
+		System.arraycopy(_data, 0, ret, this._size, other.size());
+		return new StringArray(ret);
+	}
+
+	@Override
 	public void write(DataOutput out) throws IOException {
 		out.writeByte(FrameArrayType.STRING.ordinal());
 		for(int i = 0; i < _size; i++)
@@ -380,7 +389,7 @@ public class StringArray extends Array<String> {
 	}
 
 	@Override
-	public double getAsDouble(int i){
+	public double getAsDouble(int i) {
 		return Double.parseDouble(_data[i]);
 	}