--- a/src/main/java/org/apache/sysds/runtime/frame/data/columns/StringArray.java
+++ b/src/main/java/org/apache/sysds/runtime/frame/data/columns/StringArray.java
@@ -111,7 +111,8 @@ public class StringArray extends Array<String> {
 		final int endSize = this._size + other.size();
 		final String[] ret = new String[endSize];
 		System.arraycopy(_data, 0, ret, 0, this._size);
-		System.arraycopy((String[])other.get(), 0, ret, this._size, other.size());;
+		System.arraycopy((String[]) other.get(), 0, ret, this._size, other.size());
+		;
 		return new StringArray(ret);
 	}
 
@@ -143,8 +144,11 @@ public class StringArray extends Array<String> {
 
 	@Override
 	public void reset(int size) {
-		if(_data.length < size)
+		if(_data.length < size || _data.length > 2 * size)
 			_data = new String[size];
+		else
+			for(int i = 0; i < size; i++)
+				_data[i] = null;
 		_size = size;
 	}