--- a/src/test/java/org/apache/sysds/test/component/frame/array/FrameArrayTests.java
+++ b/src/test/java/org/apache/sysds/test/component/frame/array/FrameArrayTests.java
@@ -713,11 +713,11 @@ public class FrameArrayTests {
 	public void append60Null() {
 		Array<?> aa = a.clone();
 
-		try{
+		try {
 
 			for(int i = 0; i < 60; i++)
 				aa.append((String) null);
-	
+
 			switch(a.getValueType()) {
 				case BOOLEAN:
 					assertEquals((Boolean) aa.get(aa.size() - 1), false);
@@ -745,7 +745,7 @@ public class FrameArrayTests {
 					throw new DMLRuntimeException("Invalid type");
 			}
 		}
-		catch(Exception e){
+		catch(Exception e) {
 			e.printStackTrace();
 			fail(e.getMessage());
 		}
@@ -788,6 +788,59 @@ public class FrameArrayTests {
 		}
 
 		compare(aa, a);
+	}
+
+
+	@Test
+	@SuppressWarnings("unchecked")
+	public void testSetNzString() {
+		Array<?> aa = a.clone();
+		Array<String> af = (Array<String>)aa.changeType(ValueType.STRING);
+		try{
+
+			aa.setFromOtherTypeNz(af);
+		}
+		catch(Exception e){
+			e.printStackTrace();
+			fail(e.getMessage());
+		}
+
+		compare(aa, a);
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	public void testSetFromString() {
+		Array<?> aa = a.clone();
+		Array<String> af = (Array<String>)aa.changeType(ValueType.STRING);
+		try{
+
+			aa.setFromOtherType(0, af.size()-1, af);
+		}
+		catch(Exception e){
+			e.printStackTrace();
+			fail(e.getMessage());
+		}
+
+		compare(aa, a);
+	}
+
+	@Test
+	public void resetTestCase() {
+		Array<?> aa = a.clone();
+		aa.reset(10);
+		if(aa.getValueType() == ValueType.STRING){
+			for(int i = 0; i < 10; i++) {
+				assertEquals(null, aa.get(i));
+			}
+		}
+		else{
+
+			String v = aa.get(0).toString();
+			for(int i = 1; i < 10; i++) {
+				assertEquals(v, aa.get(i).toString());
+			}
+		}
 	}
 
 	protected static void compare(Array<?> a, Array<?> b) {