Frame Cbind is implemented naively via appending rows one at a time and executing single threaded.

Current execution of a 64kx2k matrix  of:

```txt
m = read($1)
for(i in 1:$2){
    m2 = rbind(m,m)
}
print(toString(m2[1:10]))
```

```txt

SystemDS Statistics:
Total elapsed time:		50.224 sec.
Total compilation time:		0.385 sec.
Total execution time:		49.840 sec.
Cache hits (Mem/Li/WB/FS/HDFS):	11/0/0/0/1.
Cache writes (Li/WB/FS/HDFS):	0/6/0/0.
Cache times (ACQr/m, RLS, EXP):	0.122/0.002/0.002/0.000 sec.
HOP DAGs recompiled (PRED, SB):	0/0.
HOP DAGs recompile time:	0.000 sec.
Total JIT compile time:		0.459 sec.
Total JVM GC count:		0.
Total JVM GC time:		0.0 sec.
Heavy hitter instructions:
 #  Instruction  Time(s)  Count
 1  append        49.813      5
 2  toString       0.012      1
```

There are optimizations available in  parallelization and appending a block at a time, both going to be implemented in this PR.