@Baunsgaard 2023-01-05
After:

```txt
SystemDS Statistics:
Total elapsed time:		0.489 sec.
Total compilation time:		0.255 sec.
Total execution time:		0.234 sec.
Cache hits (Mem/Li/WB/FS/HDFS):	11/0/0/0/1.
Cache writes (Li/WB/FS/HDFS):	0/6/0/0.
Cache times (ACQr/m, RLS, EXP):	0.095/0.002/0.002/0.000 sec.
HOP DAGs recompiled (PRED, SB):	0/0.
HOP DAGs recompile time:	0.000 sec.
Total JIT compile time:		0.502 sec.
Total JVM GC count:		0.
Total JVM GC time:		0.0 sec.
Heavy hitter instructions:
 #  Instruction  Time(s)  Count
 1  append         0.211      5
 2  createvar      0.010      7
 3  rightIndex     0.005      1
 4  toString       0.005      1
 5  assignvar      0.001      3
```